reorganize pipelines into separate files per instance_type or behavior
